backend/requirements.txt
openai>=1.12.0
moviepy>=1.0.3
python-dotenv>=1.0.0
numpy>=1.24.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
requests>=2.31.0

